+++ src/FluentlyHttpClient/FluentHttpRequest.cs
@@ -65,6 +65,7 @@ namespace FluentlyHttpClient
 		/// <summary>
 		/// Formatters to be used for content negotiation for "Accept" and also sending formats. e.g. (JSON, XML)
 		/// </summary>
+		[Obsolete("This was added to be passed down to the middleware. Instead in middleware use FluentHttpMiddlewareClientContext.Formatters.")]
 		public MediaTypeFormatterCollection Formatters { get; set; }
 
 		/// <summary>